+++ bin/pwa.dart
@@ -135,10 +135,10 @@ Future generateWorkerScript(ArgResults argv, String libInclude) async {
   String customImport =
       'import \'package:$libInclude/pwa/offline_urls.g.dart\' as offline;';
   String createWorker =
-      'PwaWorker worker = new PwaWorker()..offlineUrls = offline.offlineUrls;';
+      'Worker worker = new Worker()..offlineUrls = offline.offlineUrls;';
   if (hasWorkerConfig) {
     customImport = 'import \'package:$libInclude/pwa/worker.dart\' as custom;';
-    createWorker = 'PwaWorker worker = custom.createWorker();';
+    createWorker = 'Worker worker = custom.createWorker();';
   }
 
   String src = '''import 'package:pwa/worker.dart';